fix: better error handling
During backup there are multiple operations that may fail which were ignored.
This commit begins to address this issue by forwarding the errors back to the main function
that can tell the user if the backup was successful or not.
Since `panic()` stops the execution of the program, by returning the error
to upper levels we can enable more sophisticated notification modes without
worring updating every single function that may fail.